+++ Modules/Inventario/Http/Controllers/ResponsablesController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Modules\Inventario\Http\Controllers;
+
+use App\Http\Controllers\Controller;
+
+class ResponsablesController extends Controller
+{
+    public function index()
+    {
+        return view('inventario::responsables.index');
+    }
+}