+++ config/versiones.php
@@ -1,9 +1,9 @@
 <?php
 
 return [
-    'BhagamApps'    => '1.8.0',
+    'BhagamApps'    => '1.9.0',
     'User'          => '2.2.1',
-    'Inventario'    => '2.6.0',
+    'Inventario'    => '2.7.0',
     'Apps'          => '1.5.0',
     'CrudGenerator' => '1.1.0',
 ];